chore(auth): reorganize auth javascript
