+++ assets/blocks/reader-registration/editor.scss
@@ -1,5 +1,5 @@
 @use './style';
-@use '../../reader-activation/auth';
+@use '../../reader-activation-auth/style' as auth;
 @use '~@wordpress/base-styles/colors' as wp-colors;
 
 .newspack-registration {
