+++ assets/reader-activation-auth/index.js
@@ -3,63 +3,12 @@
 /**
  * Internal dependencies.
  */
-import './auth.scss';
+import { domReady, formatTime } from './utils';
 
-/**
- * Specify a function to execute when the DOM is fully loaded.
- *
- * @see https://github.com/WordPress/gutenberg/blob/trunk/packages/dom-ready/
- *
- * @param {Function} callback A function to execute after the DOM is ready.
- * @return {void}
- */
-function domReady( callback ) {
-	if ( typeof document === 'undefined' ) {
-		return;
-	}
-	if (
-		document.readyState === 'complete' || // DOMContentLoaded + Images/Styles/etc loaded, so we call directly.
-		document.readyState === 'interactive' // DOMContentLoaded fires at this point, so we call directly.
-	) {
-		return void callback();
-	}
-	// DOMContentLoaded has not fired yet, delay callback until then.
-	document.addEventListener( 'DOMContentLoaded', callback );
-}
+import './google-oauth';
+import './otp-input';
 
-/**
- * Format time in MM:SS format.
- *
- * @param {number} time Time in seconds.
- */
-function formatTime( time ) {
-	const minutes = Math.floor( time / 60 );
-	const seconds = time % 60;
-	return `${ minutes }:${ seconds < 10 ? '0' : '' }${ seconds }`;
-}
-
-/**
- * Converts FormData into an object.
- *
- * @param {FormData} formData       The form data to convert.
- * @param {Array}    includedFields Form fields to include.
- *
- * @return {Object} The converted form data.
- */
-const convertFormDataToObject = ( formData, includedFields = [] ) =>
-	Array.from( formData.entries() ).reduce( ( acc, [ key, val ] ) => {
-		if ( ! includedFields.includes( key ) ) {
-			return acc;
-		}
-		if ( key.indexOf( '[]' ) > -1 ) {
-			key = key.replace( '[]', '' );
-			acc[ key ] = acc[ key ] || [];
-			acc[ key ].push( val );
-		} else {
-			acc[ key ] = val;
-		}
-		return acc;
-	}, {} );
+import './style.scss';
 
 const SIGN_IN_MODAL_HASHES = [ 'signin_modal', 'register_modal' ];
 
@@ -552,184 +501,6 @@ window.newspackRAS.push( function ( readerActivation ) {
 								.catch( () => {
 									form.endLoginFlow();
 								} );
-						}
-					} );
-			} );
-		} );
-
-		/**
-		 * OTP Input
-		 */
-		const otpInputs = document.querySelectorAll( 'input[name="otp_code"]' );
-		otpInputs.forEach( originalInput => {
-			const length = parseInt( originalInput.getAttribute( 'maxlength' ) );
-			if ( ! length ) {
-				return;
-			}
-			const inputContainer = originalInput.parentNode;
-			inputContainer.removeChild( originalInput );
-			const values = [];
-			const otpCodeInput = document.createElement( 'input' );
-			otpCodeInput.setAttribute( 'type', 'hidden' );
-			otpCodeInput.setAttribute( 'name', 'otp_code' );
-			inputContainer.appendChild( otpCodeInput );
-			for ( let i = 0; i < length; i++ ) {
-				const digit = document.createElement( 'input' );
-				digit.setAttribute( 'type', 'text' );
-				digit.setAttribute( 'maxlength', '1' );
-				digit.setAttribute( 'pattern', '[0-9]' );
-				digit.setAttribute( 'autocomplete', 'off' );
-				digit.setAttribute( 'inputmode', 'numeric' );
-				digit.setAttribute( 'data-index', i );
-				digit.addEventListener( 'keydown', ev => {
-					const prev = inputContainer.querySelector( `[data-index="${ i - 1 }"]` );
-					const next = inputContainer.querySelector( `[data-index="${ i + 1 }"]` );
-					switch ( ev.key ) {
-						case 'Backspace':
-							ev.preventDefault();
-							ev.target.value = '';
-							if ( prev ) {
-								prev.focus();
-							}
-							values[ i ] = '';
-							otpCodeInput.value = values.join( '' );
-							break;
-						case 'ArrowLeft':
-							ev.preventDefault();
-							if ( prev ) {
-								prev.focus();
-							}
-							break;
-						case 'ArrowRight':
-							ev.preventDefault();
-							if ( next ) {
-								next.focus();
-							}
-							break;
-						default:
-							if ( ev.key.match( /^[0-9]$/ ) ) {
-								ev.preventDefault();
-								ev.target.value = ev.key;
-								ev.target.dispatchEvent(
-									new Event( 'input', {
-										bubbles: true,
-										cancelable: true,
-									} )
-								);
-								if ( next ) {
-									next.focus();
-								}
-							}
-							break;
-					}
-				} );
-				digit.addEventListener( 'input', ev => {
-					if ( ev.target.value.match( /^[0-9]$/ ) ) {
-						values[ i ] = ev.target.value;
-					} else {
-						ev.target.value = '';
-					}
-					otpCodeInput.value = values.join( '' );
-				} );
-				digit.addEventListener( 'paste', ev => {
-					ev.preventDefault();
-					const paste = ( ev.clipboardData || window.clipboardData ).getData( 'text' );
-					if ( paste.length !== length ) {
-						return;
-					}
-					for ( let j = 0; j < length; j++ ) {
-						if ( paste[ j ].match( /^[0-9]$/ ) ) {
-							const digitInput = inputContainer.querySelector( `[data-index="${ j }"]` );
-							digitInput.value = paste[ j ];
-							values[ j ] = paste[ j ];
-						}
-					}
-					otpCodeInput.value = values.join( '' );
-				} );
-				inputContainer.appendChild( digit );
-			}
-		} );
-
-		/**
-		 * Third party auth.
-		 */
-		const loginsElements = document.querySelectorAll( '.newspack-reader__logins' );
-		[ ...loginsElements ].forEach( element => {
-			element.classList.remove( 'newspack-reader__logins--disabled' );
-		} );
-		const googleLoginElements = document.querySelectorAll( '.newspack-reader__logins__google' );
-		googleLoginElements.forEach( googleLoginElement => {
-			const googleLoginForm = googleLoginElement.closest( 'form' );
-			const redirectInput = googleLoginForm.querySelector( 'input[name="redirect"]' );
-			const checkLoginStatus = metadata => {
-				fetch(
-					`/wp-json/newspack/v1/login/google/register?metadata=${ JSON.stringify( metadata ) }`
-				)
-					.then( res => {
-						res
-							.json()
-							.then( ( { message, data } ) => {
-								const redirect = redirectInput?.value || null;
-								if ( googleLoginForm?.endLoginFlow ) {
-									googleLoginForm.endLoginFlow( message, res.status, data, redirect );
-								}
-							} )
-							.catch( error => {
-								if ( googleLoginForm?.endLoginFlow ) {
-									googleLoginForm.endLoginFlow( error?.message, res.status );
-								}
-							} );
-					} )
-					.catch( error => {
-						if ( googleLoginForm?.endLoginFlow ) {
-							googleLoginForm.endLoginFlow( error?.message );
-						}
-					} );
-			};
-
-			googleLoginElement.addEventListener( 'click', () => {
-				if ( googleLoginForm?.startLoginFlow ) {
-					googleLoginForm.startLoginFlow();
-				}
-
-				const metadata = googleLoginForm
-					? convertFormDataToObject( new FormData( googleLoginForm ), [ 'lists[]' ] )
-					: {};
-				metadata.current_page_url = window.location.href;
-				const authWindow = window.open(
-					'about:blank',
-					'newspack_google_login',
-					'width=500,height=600'
-				);
-				fetch( '/wp-json/newspack/v1/login/google' )
-					.then( res => res.json().then( data => Promise.resolve( { data, status: res.status } ) ) )
-					.then( ( { data, status } ) => {
-						if ( status !== 200 ) {
-							if ( authWindow ) {
-								authWindow.close();
-							}
-							if ( googleLoginForm?.endLoginFlow ) {
-								googleLoginForm.endLoginFlow( data.message, status );
-							}
-						} else if ( authWindow ) {
-							authWindow.location = data;
-							const interval = setInterval( () => {
-								if ( authWindow.closed ) {
-									checkLoginStatus( metadata );
-									clearInterval( interval );
-								}
-							}, 500 );
-						} else if ( googleLoginForm?.endLoginFlow ) {
-							googleLoginForm.endLoginFlow( newspack_reader_auth_labels.blocked_popup );
-						}
-					} )
-					.catch( error => {
-						console.log( error );
-						if ( googleLoginForm?.endLoginFlow ) {
-							googleLoginForm.endLoginFlow( error?.message, 400 );
-						}
-						if ( authWindow ) {
-							authWindow.close();
 						}
 					} );
 			} );
